app/deploy/requirements.txt
attrs==21.4.0
aws-cdk-lib==2.1.0
black==22.1.0
cattrs==1.10.0
click==8.0.4
constructs==10.0.81
isort==5.10.1
jsii==1.54.0
mypy==0.931
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.1.1
